refactor: Move Command class to separate file
Moves SystemCommands::Command to its own file for better organization:
- Create lib/octoprint/system_commands/command.rb
- Remove nested Command class from SystemCommands
- Rely on Zeitwerk autoloading instead of explicit require
- All tests still pass with full functionality

This improves code organization by keeping classes in separate files
while maintaining the same namespace structure.
